fix(bom): enable float quantities and comma separator support
- Change out_tht_quantity type from ?int to ?float to support decimals
- Add backend normalization for comma decimal separators in quantity inputs

+++ src/classes/Utils/Bom/class-bom.php
@@ -9,7 +9,7 @@ class Bom {
     public int $id;
     public int $deviceId;
     public ?int $laminateId = null;
-    public ?int $out_tht_quantity = null;
+    public ?float $out_tht_quantity = null;
     public ?string $version = null; 
     public bool $isActive;
     public string $name;
